removed new effect class and condition, added hint
