+++ Mage.Sets/src/mage/cards/k/KujaGenomeSorcerer.java
@@ -3,10 +3,15 @@ package mage.cards.k;
 import java.util.UUID;
 import mage.MageInt;
 import mage.abilities.Ability;
+import mage.abilities.condition.Condition;
 import mage.abilities.condition.common.PermanentsOnTheBattlefieldCondition;
 import mage.abilities.decorator.ConditionalOneShotEffect;
+import mage.abilities.dynamicvalue.common.PermanentsOnBattlefieldCount;
 import mage.abilities.effects.common.CreateTokenEffect;
 import mage.abilities.effects.common.TransformSourceEffect;
+import mage.abilities.hint.ConditionHint;
+import mage.abilities.hint.Hint;
+import mage.abilities.hint.ValueHint;
 import mage.abilities.keyword.TransformAbility;
 import mage.abilities.triggers.BeginningOfEndStepTriggeredAbility;
 import mage.constants.ComparisonType;
@@ -16,14 +21,16 @@ import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
 import mage.filter.common.FilterControlledPermanent;
-import mage.game.permanent.token.WizardBlackToken;
+import mage.game.permanent.token.BlackWizardToken;
 
 /**
  * @author balazskristof
  */
 public final class KujaGenomeSorcerer extends CardImpl {
 
-    private static final FilterControlledPermanent filter = new FilterControlledPermanent(SubType.WIZARD, "Wizards");
+    private static final FilterControlledPermanent filter = new FilterControlledPermanent(SubType.WIZARD);
+    private static final Condition condition = new PermanentsOnTheBattlefieldCondition(filter, ComparisonType.OR_GREATER, 4);
+    private static final Hint hint = new ValueHint("Wizards you control", new PermanentsOnBattlefieldCount(filter));
 
     public KujaGenomeSorcerer(UUID ownerId, CardSetInfo setInfo) {
         super(ownerId, setInfo, new CardType[]{CardType.CREATURE}, "{2}{B}{R}");
@@ -39,13 +46,14 @@ public final class KujaGenomeSorcerer extends CardImpl {
 
         // At the beginning of your end step, create a tapped 0/1 black Wizard creature token with "Whenever you cast a noncreature spell, this token deals 1 damage to each opponent.", Then if you control four or more Wizards, transform Kuja.
         Ability ability = new BeginningOfEndStepTriggeredAbility(
-                new CreateTokenEffect(new WizardBlackToken(), 1, true)
+                new CreateTokenEffect(new BlackWizardToken(), 1, true)
         );
         ability.addEffect(new ConditionalOneShotEffect(
                 new TransformSourceEffect(),
-                new PermanentsOnTheBattlefieldCondition(filter, ComparisonType.OR_GREATER, 4),
+                condition,
                 "if you control four or more Wizards, transform {this}"
         ).concatBy("Then"));
+        ability.addHint(hint);
         this.addAbility(ability);
         this.addAbility(new TransformAbility());
     }
